requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
loguru==0.7.2
sqlmodel==0.0.16
pydantic-settings==2.1.0
python-multipart==0.0.9
jinja2==3.1.3
aiosqlite==0.19.0
